DELFT3DFM-896: Added an extra check for the presence of an explicit working directory in case of deleting files at a WAQ-model output cleanup. Former-commit-id: ed27045942d49b724a570068d48dcbf7fc5abe45 Former-commit-id: 4c1212807ba71137a3ac9969794e0c39dbbb53fa

+++ src/DeltaShell.Plugins.DelftModels.WaterQualityModel/WaterQualityModel.cs
@@ -1251,7 +1251,7 @@ namespace DeltaShell.Plugins.DelftModels.WaterQualityModel
             var outputDirectory = ModelSettings.OutputDirectory;
             FileUtils.DeleteIfExists(outputDirectory);
             FileUtils.CreateDirectoryIfNotExists(outputDirectory);
-            filesToDeleteFromExplicitWorkingDirectoryAfterModelCleanup.ForEach(file => FileUtils.DeleteIfExists(Path.Combine(ExplicitWorkingDirectory, file)));
+            if(ExplicitWorkingDirectory != null) filesToDeleteFromExplicitWorkingDirectoryAfterModelCleanup.ForEach(file => FileUtils.DeleteIfExists(Path.Combine(ExplicitWorkingDirectory, file)));
         }
 
         # endregion
